requirements.txt
fastapi==0.112.2
fastapi-cli==0.0.5
feedparser==6.0.11
filetype==1.2.0
frozenlist==1.4.1
pydantic==2.8.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
motor==3.5.1
gnews==0.3.7
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google_search_results==2.4.2
googleapis-common-protos==1.65.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.111
pinecone-client==5.0.1                  
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7 
python-dotenv==1.0.1
python-multipart==0.0.9
pypdf==4.3.1
regex==2024.7.24
requests==2.32.3


